requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
markupsafe==2.1.3
werkzeug==2.3.7
jinja2==3.1.2